pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "osl-data-science"
version = "0.1.0"  # semantic-release
description = "Galery of Data Science Projects"
readme = "README.md"
authors = ["Ivan Ogasawara <ivan.ogasawara@gmail.com>"]
packages = [
  {include = "osl_data_science", from="src"},
]
license = "BSD 3 Clause"
exclude = [
  ".git/*",
  ".env*",
]
include = ["src/osl_data_science/py.typed"]


[tool.poetry.dependencies]
python = ">=3.9,<4"
dash = ">=2.18"
plotly = ">=5.24"
pandas = ">=2.2"
Jinja2 = ">=3.1.2"
mkdocs-include-markdown-plugin = ">=7.1"
scikit-learn = ">=1.6"
pyyaml = ">=6"
numpy = "<3"
requests = ">=2.32"
# libs for handling typing annotation
typing-extensions = ">=4.12"
eval_type_backport = {version = ">=0.2", python = "<3.10"}
types-requests = "^2.32.0.20241016"
types-pyyaml = ">=6"

[tool.poetry.group.dev.dependencies]
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.2.0"
mypy = ">=1.5.0"
bandit = ">=1.7.5"
vulture = ">=2.7"
# compose-go = ">=2.18.1"
ipython = "<8"
ipykernel = ">=6.0.0"
mkdocs = ">=1.4.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.1"
mkdocs-literate-nav = ">=0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = ">=9.1.15"
mkdocstrings = ">=0.21.2"
mkdocstrings-python = ">= 1.1.2"
mkdocs-gen-files = ">=0.5.0"
makim = "1.19.0"
# 'PosixPath' object has no attribute 'endswith'
virtualenv = "<=20.25.1"


[tool.bandit]
exclude_dirs = ["tests"]
targets = "src/osl-data-science/"
skips = ["B102", "B701", "B507", "B601"]

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
strict = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
exclude = ["scripts/"]

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
exclude = [
  'docs',
]
fix = true

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "PL",  # PL
  "RUF", # Ruff-specific rules
  "I001", # isort
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.format]
quote-style = "single"

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["./"]
sort_by_size = true
verbose = false
